Changes added - Refersh Token changes

+++ src/main/java/com/oapps/osync/rest/AuthorizeController.java
@@ -1,5 +1,6 @@
 package com.oapps.osync.rest;
 
+import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,10 +13,11 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.oapps.osync.entity.ModuleInfoEntity;
 import com.oapps.osync.entity.ServiceAuthInfoEntity;
+import com.oapps.osync.invoker.Invoker;
 import com.oapps.osync.repository.ServiceAuthInfoRepository;
 import com.oapps.osync.repository.ServiceInfoRepository;
 import com.oapps.osync.util.AuthorizeParams;
-import com.oapps.osync.util.OSyncAuthorizerUtil;
+import com.oapps.osync.util.AuthorizerUtil;
 
 import lombok.extern.java.Log;
 
@@ -37,7 +39,7 @@ public class AuthorizeController {
 		try {
 			System.out.println(" authParams >>>>>>"+authParams.toString());
 
-			String sendCodeToAuthorizationService = OSyncAuthorizerUtil.sendCodeToAuthorizationService(authParams ,serviceRepo);
+			String sendCodeToAuthorizationService = AuthorizerUtil.getAccessToken(authParams ,serviceRepo);
 
 			mapper.disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
 			AuthorizeParams newAuthParamObj = mapper.readValue(sendCodeToAuthorizationService, AuthorizeParams.class);
@@ -57,6 +59,7 @@ public class AuthorizeController {
 
 		String osyncId = decryptedArray[0];
 		String serviceId = decryptedArray[1];
+		String integId = decryptedArray[2];
 		ServiceAuthInfoEntity entityObj = new ServiceAuthInfoEntity();
 		
 		ServiceAuthInfoEntity findTopByOsyncIdAndServiceId = serviceAuthRepo.findTopByOsyncIdAndServiceId(Long.valueOf(osyncId), Long.valueOf(serviceId));
@@ -65,6 +68,7 @@ public class AuthorizeController {
 		}
 		entityObj.setAccessToken(authParams.getAccess_token());
 		entityObj.setOsyncId(Long.valueOf(osyncId));
+		entityObj.setIntegId(Long.valueOf(integId));
 		entityObj.setRefreshToken(authParams.getRefresh_token());
 		entityObj.setTokenType("org");
 		entityObj.setServiceId(Long.valueOf(serviceId));
@@ -74,8 +78,10 @@ public class AuthorizeController {
 	}
 	@DeleteMapping(path = "/api/v1/revoke")
 	public void revoke(@RequestParam String serviceId,@RequestParam String osyncId, @RequestParam String integId) {
-		if(serviceId != null && osyncId != null) {
+		if(serviceId != null && osyncId != null && integId != null ) {
 			serviceAuthRepo.delete(serviceAuthRepo.findByOsyncIdAndServiceIdAndIntegId(Long.valueOf(osyncId),Long.valueOf(serviceId),Long.valueOf(integId) ));
 		}
-	} 
+	}
+	
+	
 }
